+++ playground/examples/index.ts
@@ -13,6 +13,7 @@ import relative from './relative'
 import sections from './sections'
 import styling_inlines from './styling_inlines'
 import styling_named_styles from './styling_named_styles'
+import styling_named_styles_with_extends from './styling_named_styles_with_extends'
 import styling_named_styles_with_overrides from './styling_named_styles_with_overrides'
 import styling_properties from './styling_properties'
 import svgs from './svgs'
@@ -79,6 +80,10 @@ export const examples: Example[] = [
   {
     name: 'Styling (Inlines)',
     code: styling_inlines,
+  },
+  {
+    name: 'Styling (Extends)',
+    code: styling_named_styles_with_extends,
   },
   {
     name: 'Styling (Overrides)',