+++ src/components/ContentSvgRenderer.svelte
@@ -8,7 +8,15 @@
 
   $: src =
     typeof node.svg === 'string'
-      ? 'data:image/svg+xml,' + encodeURIComponent(node.svg)
+      ? 'data:image/svg+xml,' +
+        encodeURIComponent(
+          node.svg.indexOf('xmlns') === -1
+            ? node.svg.replace(
+                '<svg',
+                '<svg xmlns="http://www.w3.org/2000/svg"'
+              )
+            : node.svg
+        )
       : undefined
 
   $: if (